サンプルの見直しとバグの修正
