--- examples/mode_notification/custom_cushy_web_server.hpp
@@ -1,24 +1,0 @@
-/**
- * @file custom_cushy_web_server.hpp
- * @brief CushyWebServer を継承し、HTMLルートハンドラなどを追加したカスタムWebサーバークラスの定義
- * @version 0.3.0
- * @date 2023-03-28
- *
- * @copyright Copyright (c) 2023 / MaSiRo Project.
- *
- */
-#ifndef CUSTOM_CUSHY_WEB_SERVER_HPP
-#define CUSTOM_CUSHY_WEB_SERVER_HPP
-
-#include <cushy_web_server.hpp>
-
-class CustomCushyWebServer : public CushyWebServer {
-public:
-protected:
-    bool setup_server(AsyncWebServer *server) override;
-
-private:
-    void handle_html_root(AsyncWebServerRequest *request);
-};
-
-#endif
