--- examples/original_favicon/custom_cushy_web_server.hpp
@@ -1,25 +1,0 @@
-/**
- * @file custom_cushy_web_server.hpp
- * @brief 派生クラスでWebサーバーのセットアップとfavicon、HTMLルートの処理をカスタマイズするヘッダ
- * @version 0.3.0
- * @date 2023-03-28
- *
- * @copyright Copyright (c) 2023 / MaSiRo Project.
- *
- */
-#ifndef CUSTOM_CUSHY_WEB_SERVER_HPP
-#define CUSTOM_CUSHY_WEB_SERVER_HPP
-
-#include <cushy_web_server.hpp>
-
-class CustomCushyWebServer : public CushyWebServer {
-public:
-protected:
-    bool setup_server(AsyncWebServer *server) override;
-    void handle_favicon_ico(AsyncWebServerRequest *request) override;
-
-private:
-    void handle_html_root(AsyncWebServerRequest *request);
-};
-
-#endif
