+++ library.json
@@ -33,38 +33,22 @@
             ]
         },
         {
-            "name": "check_stack_size",
-            "base": "examples/check_stack_size",
-            "files": [
-                "check_stack_size.cpp"
-            ]
-        },
-        {
-            "name": "mode_notification",
-            "base": "examples/mode_notification",
+            "name": "standard",
+            "base": "examples/standard",
             "files": [
                 "custom_cushy_web_server.cpp",
                 "custom_cushy_web_server.hpp",
-                "mode_notification.cpp"
+                "standard.cpp"
             ]
         },
         {
-            "name": "original_favicon",
-            "base": "examples/original_favicon",
+            "name": "advance",
+            "base": "examples/advance",
             "files": [
                 "custom_cushy_web_server.cpp",
                 "custom_cushy_web_server.hpp",
                 "data_custom_cushy_web_server.hpp",
-                "original_favicon.cpp"
-            ]
-        },
-        {
-            "name": "post_json",
-            "base": "examples/post_json",
-            "files": [
-                "custom_cushy_web_server.cpp",
-                "custom_cushy_web_server.hpp",
-                "post_json.cpp"
+                "advance.cpp"
             ]
         }
     ],
