+++ platformio.ini
@@ -16,6 +16,7 @@ src_dir = ./
 monitor_speed = 115200
 build_flags =
 	-D ELEGANTOTA_USE_ASYNC_WEBSERVER=1
+	-D CORE_DEBUG_LEVEL=ARDUHAL_LOG_LEVEL_INFO
 build_src_filter =
 	+<src/>
 	-<.git/>
@@ -25,13 +26,8 @@ build_src_filter =
 	-<doxygen/>
 	-<test/>
 	-<examples/basic>
-	-<examples/check_stack_size>
-	-<examples/delete_file>
-	-<examples/list_reconnect>
-	+<examples/list_settings>
-	-<examples/mode_notification>
-	-<examples/original_favicon>
-	-<examples/post_json>
+	+<examples/standard>
+	-<examples/advance>
 
 [env:unit_build]
 framework = arduino
